Minor ConcurrencyLimiter tweaks
* Use ArrayBlockingQueue, which has more suitable semantics
* Add interface
* Track remaining capacity, not used capacity, since the
  former is more useful for monitoring purposes.
