+++ src/test/java/com/spotify/futures/jdk7/SimpleConcurrencyLimiterTest.java
@@ -21,7 +21,7 @@ import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.SettableFuture;
 
-import com.spotify.futures.ConcurrencyLimiter;
+import com.spotify.futures.SimpleConcurrencyLimiter;
 import com.spotify.futures.FuturesExtra;
 
 import org.junit.Test;
@@ -33,27 +33,27 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.junit.Assert.*;
 
-public class ConcurrencyLimiterTest {
+public class SimpleConcurrencyLimiterTest {
 
   @Test(expected = IllegalArgumentException.class)
   public void testTooLowConcurrency() throws Exception {
-    ConcurrencyLimiter.create(0, 10);
+    SimpleConcurrencyLimiter.create(0, 10);
   }
 
   @Test(expected = IllegalArgumentException.class)
   public void testTooLowQueueSize() throws Exception {
-    ConcurrencyLimiter.create(10, 0);
+    SimpleConcurrencyLimiter.create(10, 0);
   }
 
   @Test(expected = NullPointerException.class)
   public void testNullJob() throws Exception {
-    final ConcurrencyLimiter<String> limiter = ConcurrencyLimiter.create(1, 10);
+    final SimpleConcurrencyLimiter<String> limiter = SimpleConcurrencyLimiter.create(1, 10);
     limiter.add(null);
   }
 
   @Test
   public void testJobReturnsNull() throws Exception {
-    final ConcurrencyLimiter<String> limiter = ConcurrencyLimiter.create(1, 10);
+    final SimpleConcurrencyLimiter<String> limiter = SimpleConcurrencyLimiter.create(1, 10);
     final ListenableFuture<String> response = limiter.add(job(null));
     assertTrue(response.isDone());
     final Throwable exception = FuturesExtra.getException(response);
@@ -62,7 +62,7 @@ public class ConcurrencyLimiterTest {
 
   @Test
   public void testJobThrows() throws Exception {
-    final ConcurrencyLimiter<String> limiter = ConcurrencyLimiter.create(1, 10);
+    final SimpleConcurrencyLimiter<String> limiter = SimpleConcurrencyLimiter.create(1, 10);
     final ListenableFuture<String> response = limiter.add(new Callable<ListenableFuture<String>>() {
       @Override
       public ListenableFuture<String> call() throws Exception {
@@ -77,7 +77,7 @@ public class ConcurrencyLimiterTest {
 
   @Test
   public void testJobReturnsFailure() throws Exception {
-    final ConcurrencyLimiter<String> limiter = ConcurrencyLimiter.create(1, 10);
+    final SimpleConcurrencyLimiter<String> limiter = SimpleConcurrencyLimiter.create(1, 10);
     final ListenableFuture<String> response = limiter.add(job(Futures.<String>immediateFailedFuture(new IllegalStateException())));
 
     assertTrue(response.isDone());
@@ -87,7 +87,7 @@ public class ConcurrencyLimiterTest {
 
   @Test
   public void testCancellation() throws Exception {
-    final ConcurrencyLimiter<String> limiter = ConcurrencyLimiter.create(2, 10);
+    final SimpleConcurrencyLimiter<String> limiter = SimpleConcurrencyLimiter.create(2, 10);
     final SettableFuture<String> request1 = SettableFuture.create();
     final SettableFuture<String> request2 = SettableFuture.create();
 
@@ -110,31 +110,31 @@ public class ConcurrencyLimiterTest {
     assertFalse(response1.isDone());
     assertFalse(response2.isDone());
     assertTrue(response3.isDone());
-    assertEquals(2, limiter.numActive());
-    assertEquals(1, limiter.numQueued());
+    assertEquals(0, limiter.remainingActiveCapacity());
+    assertEquals(9, limiter.remainingQueueCapacity());
 
     request2.set("2");
 
     assertFalse(response1.isDone());
     assertTrue(response2.isDone());
     assertTrue(response3.isDone());
-    assertEquals(1, limiter.numActive());
-    assertEquals(0, limiter.numQueued());
+    assertEquals(1, limiter.remainingActiveCapacity());
+    assertEquals(10, limiter.remainingQueueCapacity());
 
     request1.set("1");
 
     assertTrue(response1.isDone());
     assertTrue(response2.isDone());
     assertTrue(response3.isDone());
-    assertEquals(0, limiter.numActive());
-    assertEquals(0, limiter.numQueued());
+    assertEquals(2, limiter.remainingActiveCapacity());
+    assertEquals(10, limiter.remainingQueueCapacity());
 
     assertFalse(wasInvoked.get());
   }
 
   @Test
   public void testSimple() throws Exception {
-    final ConcurrencyLimiter<String> limiter = ConcurrencyLimiter.create(2, 10);
+    final SimpleConcurrencyLimiter<String> limiter = SimpleConcurrencyLimiter.create(2, 10);
     final SettableFuture<String> request1 = SettableFuture.create();
     final SettableFuture<String> request2 = SettableFuture.create();
     final SettableFuture<String> request3 = SettableFuture.create();
@@ -149,32 +149,31 @@ public class ConcurrencyLimiterTest {
     assertFalse(response1.isDone());
     assertFalse(response2.isDone());
     assertFalse(response3.isDone());
-    assertEquals(2, limiter.numActive());
-    assertEquals(1, limiter.numQueued());
+    assertEquals(0, limiter.remainingActiveCapacity());
+    assertEquals(9, limiter.remainingQueueCapacity());
 
     request2.set("2");
 
     assertFalse(response1.isDone());
     assertTrue(response2.isDone());
     assertTrue(response3.isDone());
-    assertEquals(1, limiter.numActive());
-    assertEquals(0, limiter.numQueued());
+    assertEquals(1, limiter.remainingActiveCapacity());
+    assertEquals(10, limiter.remainingQueueCapacity());
 
     request1.set("1");
 
     assertTrue(response1.isDone());
     assertTrue(response2.isDone());
     assertTrue(response3.isDone());
-    assertEquals(0, limiter.numActive());
-    assertEquals(0, limiter.numQueued());
-
+    assertEquals(2, limiter.remainingActiveCapacity());
+    assertEquals(10, limiter.remainingQueueCapacity());
   }
 
   @Test
   public void testLongRunning() throws Exception {
     final AtomicInteger activeCount = new AtomicInteger();
     final AtomicInteger maxCount = new AtomicInteger();
-    final ConcurrencyLimiter<String> limiter = ConcurrencyLimiter.create(10, 100000);
+    final SimpleConcurrencyLimiter<String> limiter = SimpleConcurrencyLimiter.create(10, 100000);
     List<CountingJob> jobs = Lists.newArrayList();
     List<ListenableFuture<String>> responses = Lists.newArrayList();
     for (int i = 0; i < 100000; i++) {
@@ -197,13 +196,13 @@ public class ConcurrencyLimiterTest {
     }
     assertEquals(10, maxCount.get());
     assertEquals(0, activeCount.get());
-    assertEquals(0, limiter.numActive());
-    assertEquals(0, limiter.numQueued());
+    assertEquals(10, limiter.remainingActiveCapacity());
+    assertEquals(100000, limiter.remainingQueueCapacity());
   }
 
   @Test
   public void testQueueSize() throws Exception {
-    final ConcurrencyLimiter<String> limiter = ConcurrencyLimiter.create(10, 10);
+    final SimpleConcurrencyLimiter<String> limiter = SimpleConcurrencyLimiter.create(10, 10);
     for (int i = 0; i < 20; i++) {
       limiter.add(job(SettableFuture.<String>create()));
     }
@@ -211,25 +210,25 @@ public class ConcurrencyLimiterTest {
     final ListenableFuture<String> future = limiter.add(job(SettableFuture.<String>create()));
     assertTrue(future.isDone());
     final Throwable e = FuturesExtra.getException(future);
-    assertEquals(ConcurrencyLimiter.CapacityReachedException.class, e.getClass());
+    assertEquals(SimpleConcurrencyLimiter.CapacityReachedException.class, e.getClass());
   }
 
   @Test
   public void testQueueSizeCounter() throws Exception {
     final SettableFuture<String> future = SettableFuture.create();
 
-    final ConcurrencyLimiter<String> limiter = ConcurrencyLimiter.create(10, 10);
+    final SimpleConcurrencyLimiter<String> limiter = SimpleConcurrencyLimiter.create(10, 10);
     for (int i = 0; i < 20; i++) {
       limiter.add(job(future));
     }
 
-    assertEquals(10, limiter.numActive());
-    assertEquals(10, limiter.numQueued());
+    assertEquals(0, limiter.remainingActiveCapacity());
+    assertEquals(0, limiter.remainingQueueCapacity());
 
     future.set("");
 
-    assertEquals(0, limiter.numActive());
-    assertEquals(0, limiter.numQueued());
+    assertEquals(10, limiter.remainingActiveCapacity());
+    assertEquals(10, limiter.remainingQueueCapacity());
   }
 
   private Callable<ListenableFuture<String>> job(final ListenableFuture<String> future) {
